@@ -1,19 +1,23 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <LangVersion>8</LangVersion>
+    <TargetFrameworks>net8.0;netcoreapp3.1;netstandard2.1;net481</TargetFrameworks>
+    <LangVersion>12</LangVersion>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Description>A simple implementation of round robin in c# with circular linked list</Description>
+    <Description>This Package provides Round Robin algorithm that is thread safe, It can be used for load balancing </Description>
     <Copyright>None</Copyright>
     <PackageProjectUrl>https://github.com/alicommit-malp/roundrobin</PackageProjectUrl>
     <RepositoryUrl>https://github.com/alicommit-malp/roundrobin</RepositoryUrl>
     <RepositoryType>GIT</RepositoryType>
-    <PackageTags>LinkedList , RoundRobin, Circular Linked List</PackageTags>
-    <Title>Round Robin thread safe algorithm with linked list </Title>
+    <PackageTags>Round Robin,Roundrobin, Thread Safe,thread-safe,Load balancing</PackageTags>
+    <Title>Round Robin (thread safe)</Title>
     <license>MIT</license>
-    <PackageVersion>2.2.1</PackageVersion>
+    <PackageVersion>2.3.0</PackageVersion>
     <LicenseUrl>https://raw.githubusercontent.com/alicommit-malp/roundrobin/master/LICENSE</LicenseUrl>
+  </PropertyGroup>
+
+  <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
+    <DocumentationFile>bin\Debug\RoundRobin.xml</DocumentationFile>
   </PropertyGroup>
 
 </Project>
